Speedup SpaceToDepth on CPU. Thread over batch and memcpy over channel. Speedup 1.1x ~ 3x for large number of batches or channels.

On **Intel(R) Core(TM) i5-9600K CPU @ 3.70GHz**

| Input shape `(N x H x W x C)` | Block size | Original (M items / s) | This PR (M items / s) |
| --------------------------------| -----------| ----------| ---------|
| 1 x 64 x 64 x 1                        | 2              |   83            |   73 |
| 16 x 64 x 64 x 1                        | 2              |  241              |  337 |
| 32 x 64 x 64 x 1                        | 2              |   261            |  559 |
| 1 x 128 x 128x 1                        | 2              |   162            |   132 |
| 16 x 128 x 128 x 1                        | 2              | 308              |   772 |
| 32 x 128 x 128 x 1                        | 2              |   318            |    931 |
| 1 x 64 x 64 x 3                        | 2              |     238          |   221 |
| 16 x 64 x 64 x 3                        | 2              |   617            | 1059 |
| 32 x 64 x 64 x 3                        | 2              |    665           |  1609 |
| 1 x 128 x 128x 3                        | 2              |  437             | 391 |
| 16 x 128 x 128 x 3                        | 2              | 737              |   2146 |
| 32 x 128 x 128 x 3                        | 2              |   754            |   2221 |
| 1 x 64 x 64 x 15                        | 2              |     691           |   841 |
| 16 x 64 x 64 x 15                        | 2              |   1375            |   2752 |
| 32 x 64 x 64 x 15                        | 2              |    1369           |  2024 |
| 1 x 128 x 128x 15                        | 2              |   1216            |  1770 |
| 16 x 128 x 128 x 15                        | 2              |   1248            |    2159 |
| 32 x 128 x 128 x 15                        | 2              |    415           |    2062 |
| 1 x 64 x 64 x 64                        | 2              |  1455             |   2590 |
| 16 x 64 x 64 x 64                        | 2              |  1478             |  2165 |
| 32 x 64 x 64 x 64                        | 2              |   1259            |   1658 |
| 1 x 128 x 128x 64                        | 2              |   1725            |   2080 |
| 16 x 128 x 128 x 64                        | 2              |  1255             |   1552 |
| 32 x 128 x 128 x 64                        | 2              |  1209             |    1644 |
| 1 x 64 x 64 x 64                        | 4              |   1541            |   2698 |
| 16 x 64 x 64 x 64                        | 4              | 1602              |   2136 |
| 32 x 64 x 64 x 64                        | 4              |  1346             |   1616 |
| 1 x 128 x 128x 64                        | 4              |   1618            |   1969 |
| 16 x 128 x 128 x 64                        | 4              |  1252             |   1579 |
| 32 x 128 x 128 x 64                        | 4              |  1243             |   1656 |

where the number is computed as `time / (iterations * N * H * W * C)`.